refactor(app): move a module prep command provider
