+++ app/src/local-resources/modules/index.ts
@@ -1,0 +1,1 @@
+export * from './getModulePrepCommands'
