+++ app/src/organisms/ModuleCard/index.tsx
@@ -60,7 +60,7 @@ import { HeaterShakerModuleData } from './HeaterShakerModuleData'
 import { HeaterShakerSlideout } from './HeaterShakerSlideout'
 import { TestShakeSlideout } from './TestShakeSlideout'
 import { ModuleWizardFlows } from '../ModuleWizardFlows'
-import { getModulePrepCommands } from '/app/organisms/Devices/getModulePrepCommands'
+import { getModulePrepCommands } from '/app/local-resources/modules'
 import { getModuleCardImage } from './utils'
 import { FirmwareUpdateFailedModal } from './FirmwareUpdateFailedModal'
 import { ErrorInfo } from './ErrorInfo'
